+++ tests/test_commands.py
@@ -237,7 +237,7 @@ class NewCommandTests(CommandTestCase):
     def test_create_issue_error(self):
         self.server.set_error_response(400, "Field 'priority' is required")
 
-        result = self.invoke('create', 'GOJI', 'Sprint #1', 'x', '--description', 'Desc')
+        result = self.invoke('create', 'GOJI', 'Sprint #1', '--description', 'Desc')
 
         self.assertEqual(result.output, "Description:\n\nDesc\n\nField 'priority' is required\n")
         self.assertEqual(result.exit_code, 1)
@@ -245,7 +245,13 @@ class NewCommandTests(CommandTestCase):
     def test_new_title_description(self):
         self.server.set_create_issue_response()
 
-        result = self.invoke('create', 'GOJI', 'Sprint #1', 'x', '--description', 'Desc')
+        result = self.invoke('create', 'GOJI', 'Sprint #1', '--description', 'Desc')
+
+        self.assertEqual(self.server.last_request.body, {'fields': {
+            'description': 'Desc',
+            'project': {'key': 'GOJI'},
+            'summary': 'Sprint #1',
+        }})
 
         self.assertIsNone(result.exception)
         self.assertEqual(result.output, 'Issue GOJI-133 created\n')
@@ -258,7 +264,20 @@ class NewCommandTests(CommandTestCase):
         pass
 
     def test_new_specify_type(self):
-        pass
+        self.server.set_create_issue_response()
+
+        result = self.invoke('create', 'GOJI', 'Sprint #1', '--type', 'Bug', '--description', 'Desc')
+
+        self.assertEqual(self.server.last_request.body, {'fields': {
+            'description': 'Desc',
+            'issuetype': {'name': 'Bug'},
+            'project': {'key': 'GOJI'},
+            'summary': 'Sprint #1',
+        }})
+
+        self.assertIsNone(result.exception)
+        self.assertEqual(result.output, 'Issue GOJI-133 created\n')
+        self.assertEqual(result.exit_code, 0)
 
 
 class CreateSprintTests(CommandTestCase):